==> src/02-Plots.qmd <==
# Plots

```{r}
# Load necessary libraries
library(tidyverse)
library(ggthemes)

df <- read_csv("../data/DWA_ECB_clean.csv")
df_simple <- read_csv("../data/DWA_ECB_simplified.csv")
```


### Tenant - Owner Wealth Ratio

```{r}
df_ratio <- df %>%
  filter(
    INSTR_ASSET == "NWA",
    DWA_GRP %in% c("HSO", "HST"),
    UNIT_MEASURE == "EUR_R_POP"
  ) %>%
  select(REF_AREA, DWA_GRP, OBS_VALUE, TIME_PERIOD) %>%
  pivot_wider(
    names_from = DWA_GRP,
    values_from = OBS_VALUE
  ) %>%
  mutate(
    ratio = HSO / HST
  ) 


p2 <- df_ratio %>%
  filter(REF_AREA %in% c("DE", "FR", "IT", "ES", "NL", "BE", "GR", "HU")) %>%
  ggplot(aes(x = TIME_PERIOD, y = ratio, color = REF_AREA)) +
    geom_line(linewidth=0.5) +
    geom_point(shape=5, size= 1) 
p2
```

```{r}
p2 <- p2 +
    scale_color_brewer(palette = "Set1")+  
    labs(
      x = "Time Period",
      y = "Ratio",
      color = "Country",
      caption = "Source: ECB DWA, own calculations and visualization"
    ) +
    theme_tufte() +
    theme(
      legend.position = "right",
      legend.title = element_blank(),
      panel.border = element_rect(fill = NA, color = "black"),
      ) 

ggsave(p2, filename = "../output/tenant_owner_wealth_ratio.png", width = 8, height = 5, dpi = 500)
p2
```

### Country: Asset Distribution

```{r}
df_filter <- df %>%
  filter(
    #REF_AREA == "DE",
    INSTR_ASSET %in% c("F51M", "NUB", "F62", "F52", "F511", "F3", "NUN", "F2M", "NWA", "F4X", "F4B"),
    DWA_GRP %in% c("B50", "D06", "D07", "D08", "D09", "D10")
    ) %>%
  select(-c(KEY,TITLE, COMMENT_TS)) %>%
  pivot_wider(
    names_from = INSTR_ASSET,
    values_from = OBS_VALUE
  ) %>%
  mutate(
    BW = NUB + F51M, # business wealth
    FW = F3 + F511 + F52 + F62, # financial wealth
    DEP = F2M, # deposits
    HW = NUN, # housing wealth
    HW_NET = NUN + F4B, # housing net wealth
    TW = BW + FW + HW + DEP,
    DEBT = F4X + F4B, # total debt
    TW_NET = BW + FW + DEP + HW + DEBT, # total wealth net
  ) %>%
  select(
    TIME_PERIOD, REF_AREA, DWA_GRP, UNIT_MEASURE,
    BW, FW, DEP, HW, HW_NET, TW, DEBT, TW_NET
  )
```


```{r}
df_asset <- df_filter %>%
  filter(
    TIME_PERIOD == "2023-01-01", #only one data and latest for all countrie s(NL doesnt have 2024)
    UNIT_MEASURE == "EUR" # aggregate EUR
  ) %>%
  arrange(desc(DWA_GRP)) %>% #sort GRP descending 
  # shares for all assets
  summarise(
    REF_AREA = REF_AREA,
    DWA_GRP = DWA_GRP,
    share_BW = BW / TW,
    share_FW = FW / TW,
    share_DEP = DEP / TW,
    share_HW = HW / TW
  ) %>%
  # long dataframe = easier for ploting
  pivot_longer(
    cols = starts_with("share_"),
    names_to = "asset_class",
    values_to = "share"
  ) 
p3 <- df_asset %>%
  filter(REF_AREA %in% c("DE", "IT","NL", "AT", "ES", "FR")) %>% 
  transform(REF_AREA = factor( # custom sorting of the countries (low ownership up)
    REF_AREA, levels = c("DE", "AT","NL", "IT", "ES", "FR")
    )) %>%
  ggplot(aes(x = DWA_GRP, y = share, fill = asset_class)) +
    geom_col(position = "stack") +  # stack the bars to make 1
    facet_wrap(~ REF_AREA, ncol = 3, scales = "free_x")  #x axis should appear at every plot
p3
```

```{r}
p3 <- p3 +
    scale_fill_manual(
      # change the colors of the stacks
      values = c(
        "share_BW" = "#cdb04b",
        "share_FW" = "#377EB8",
        "share_DEP" = "#38ad68",
        "share_HW" = "#9f464f"
      ),
      # change the labels
      labels = c(
        "share_BW" = "Business Wealth",
        "share_FW" = "Financial Wealth",
        "share_DEP" = "Deposits",
        "share_HW" = "Housing Wealth"
      )
    ) +
    labs(
      x = "Wealth Decile",
      y = "Share of Total Wealth",
      fill = "Asset Class",
      # special caption to explain the GRP abbreviations
      caption = "
      B50 = Bottom 50% \ \  D06 = 6th Decile \ \  D07 = 7th Decile
D08 = 8th Decile \ \ D09 = 9th Decile \ \  D10 = Top 10%
Source: ECB DWA, own calculations and visualization" 
    ) +
    theme_tufte() +
    theme(
      legend.position = "right",
      legend.title = element_blank(),
      panel.border = element_rect(fill = NA, color = "black"),
    )  
    
ggsave(p3, filename = "../output/asset_distribution_classes.png", width = 8, height = 5, dpi = 500)
p3
```


### Time: Asset Distribution

```{r}
df_growth <- df_filter %>%
  filter(
    UNIT_MEASURE == "EUR_R_POP", # only per capita
    DWA_GRP %in% c("B50", "D06", "D07", "D08", "D09", "D10") #drop other groups
  ) %>%
  select(TIME_PERIOD, REF_AREA, DWA_GRP, UNIT_MEASURE, DEP, FW, BW, HW, DEBT, TW_NET) %>%
  pivot_longer(
    cols = c(DEP, FW, BW, HW, DEBT, TW_NET),
    names_to = "INSTR_ASSET",
    values_to = "OBS_VALUE"
  )

df_growth <- df_growth %>%
  # # create a mean of all groups as a new GRP = ALL
  # group_by(TIME_PERIOD, REF_AREA, UNIT_MEASURE, INSTR_ASSET) %>%
  # summarise(OBS_VALUE = mean(OBS_VALUE, na.rm = TRUE)) %>%
  # mutate(DWA_GRP = "ALL") %>%
  # # bind the new group to the original data
  # bind_rows(
  #   df_growth %>%
  #     filter(
  #       UNIT_MEASURE == "EUR_R_POP",
  #       DWA_GRP %in% c("B50", "D06", "D07", "D08", "D09", "D10")
  #     ) %>%
  #     group_by(TIME_PERIOD, REF_AREA, DWA_GRP, UNIT_MEASURE, INSTR_ASSET) %>%
  #     summarise(OBS_VALUE = mean(OBS_VALUE, na.rm = TRUE))
  # ) %>%
  # ungroup() %>%
  # group D06 till D09 to middle class
  mutate( # create the middle class GRP
    DWA_GRP = case_when(
      DWA_GRP %in% c("D06", "D07", "D08", "D09") ~ "M40",
      TRUE ~ DWA_GRP
    )
  ) %>%
  # mean of the values of M40
  group_by(TIME_PERIOD, REF_AREA, DWA_GRP, UNIT_MEASURE, INSTR_ASSET) %>%
  summarise(OBS_VALUE = mean(OBS_VALUE, na.rm = TRUE))

df_growth %>%
  filter(
    REF_AREA == "DE",
    TIME_PERIOD == "2011-01-01",
    INSTR_ASSET == "TW_NET"
  )
```

```{r}
p4 <- df_growth %>%
  filter(
    REF_AREA %in% c("I9"),    #Euro 20 Area
    str_detect(TIME_PERIOD, "01-01|06-01"), # filter for first and third quarter
    ) %>%
  transform(DWA_GRP = factor(
    DWA_GRP, levels = c("B50","M40", "D10"), # order of the GRP for the plot
    labels = c("Bottom 50%", "Middle 40%", "Top 10%") # name of the GRP for the plot
    )) %>%
  ggplot(aes(x = TIME_PERIOD)) + #same x axis for all groups
  geom_area(
    data = ~ .x %>% filter(INSTR_ASSET != "TW_NET"), # dont add the Total to the stacks
    aes(y = OBS_VALUE, fill = INSTR_ASSET),
    position = "stack" #stack the lines on top of each other
  ) +
  # line for total
  geom_line(
    data = ~ .x %>% filter(INSTR_ASSET == "TW_NET"),
    aes(y = OBS_VALUE, group = 1, color = "Net Wealth"), #name of the line
    size = 1,
    linetype = "11" # small dots
  ) +
  facet_wrap(~ DWA_GRP, ncol = 3, scales = "free_y")  # seperate plot for all GRP, with differnt y axis

p4
```

```{r}
p4 <- p4 +
  geom_hline(
    yintercept = 0,
    color = "black",
    linewidth = 0.5
  ) +
  scale_fill_manual(
    #change the colors of the stacks
    values = c(   
        "BW" = "#cdb04b",
        "FW" = "#377EB8",
        "DEP" = "#38ad68",
        "HW" = "#9f464f",
        "DEBT" = "grey"
      ),
    # change the labels
    labels = c(
      "BW" = "Business Wealth",
      "FW" = "Financial Wealth",
      "DEP" = "Deposits",
      "HW" = "Housing Wealth",
      "DEBT" = "Debt"
    )
  ) +
  # adjust the line plot
  scale_color_manual(
    values = c("Net Wealth" = "black"),
    labels = c("Net Wealth" = "Net Wealth") 
  ) +
  labs(
    x = "Time Period",
    y = "Wealth per Capita (TEUR)",
    fill = "Asset Class",
    caption = "Source: ECB DWA, own calculations and visualization"
  ) +
  theme_tufte() +
  theme(
    legend.position = "bottom",
    legend.title = element_blank(),
    panel.border = element_rect(fill = NA, color = "black"),
  )

ggsave(p4, filename = "../output/asset_distribution_over_time.png", width = 8, height = 5, dpi = 500)
p4
```

## Ownership

```{r}
ownership <- read_csv("../data/ownership.csv")
```
```{r}
#plot Europe Differences between years
df_owner %>%
  filter(
    DWA_GRP != "ALL", # drop the average group
    REF_AREA %in% c("DE", "BE", "NL", "AT", "ES", "FR", "IT","I9")
  ) %>%
  ggplot(aes(x = DWA_GRP, y = OWNER, group=JAHR)) +
    geom_line(aes(color = as.factor(JAHR)), size = 1) +
    facet_wrap(~REF_AREA, scales = "free_x", ncol=4)
```

- relevant = I9 = Euro Area = almost no change
- different kinks, compare (AT,DE,FR) to (ES,IT,NL)

### House Prices

```{r}
df_house <- read_csv("../data/House_Prices.csv")
```


```{r}
#quick plot
p5 <- df_house %>%
  filter(
    REF_AREA %in% c("DE", "FR", "IT", "ES", "NL", "BE", "GR", "HU"), #selected
    # REF_AREA %in% c("AT", "BE", "DE", "FR", "LU", "NL", "FI", "IE"), # northern
    # REF_AREA %in% c("ES", "IT", "PT", "GR", "MT", "CY"), # southern europe
    # REF_AREA %in% c("EE", "HU", "HR", "LT", "LV", "SI", "SK"), # eastern europe
    TIME_PERIOD >= "2008-01-01"
  ) %>%
  ggplot(aes(x = TIME_PERIOD, y = HP_R_N, color = REF_AREA)) +
    geom_line(linewidth = 0.5) +
    geom_point(shape = 5, size = 0.8) +
    geom_hline(yintercept = 100, linetype = "dashed", color = "black")  # 2010 = 100
p5
```
trajectories: 
- IT, ES, = lower
- FR, BE = no change
- DE, AT, EE = increase (EE completely bonkers)
- HU, NL = down till 2015, increase after
- overall Euro ARea = increase 

maybe split up after euroepan areas (south, north, east?)

```{r}
p5 <- p5 +
    scale_color_brewer(palette = "Set1") +  
    labs(
      x = "Time Period",
      y = "House Price Index (2010 = 100)",
      color = "Country",
      caption = "Source: ECB DWA, own calculations and visualization"
    ) +
    theme_tufte() +
    theme(
      legend.position = "right",
      legend.title = element_blank(),
      panel.border = element_rect(fill = NA, color = "black"),
    )
ggsave(p5, filename = "../output/house_prices.png", width = 8, height = 5, dpi = 500)
p5
```

# Unused plots

## Gini / Theil in Europe

```{r}
df_gini <- read_csv("../data/gini_indices.csv")
```

```{r}
df_gini %>%
  filter(
    ASSET %in% c("NUN"), #housing asset
    REF_AREA %in% c("DE", "IT","NL", "AT", "ES", "FR", "BE", "PT") # selected countries
    ) %>%
  ggplot(aes(x = TIME_PERIOD)) +
    geom_line(aes(y = GINI, color = REF_AREA)) 
```

## P10 P50 ratios

```{r}
df_filter %>%
  filter(
    REF_AREA %in% c("DE", "BE", "IT", "AT", "ES", "FR"),
    DWA_GRP %in% c("B50", "D10"),
    UNIT_MEASURE == "EUR_R_POP"
  ) %>%
  select(TIME_PERIOD, REF_AREA, DWA_GRP, TW_NET) %>%
  group_by(TIME_PERIOD, REF_AREA) %>%
  summarise(
    ratio = TW_NET[DWA_GRP == "D10"] / TW_NET[DWA_GRP == "B50"]
  ) %>%
  ggplot(aes(x = TIME_PERIOD, y = ratio, color=REF_AREA)) +
    geom_line()
```
## Covariance per Decile across Europe

```{r}
df_simple %>%
  filter(
    TIME_PERIOD == "2023-01-01",
    UNIT_MEASURE == "EUR_R_POP",
    DWA_GRP %in% c("B50", "D06", "D07", "D08", "D09", "D10"),
  ) %>%
  select(REF_AREA, DWA_GRP, TW_NET) %>%
  # variance in between the countries
  group_by(DWA_GRP) %>%
  summarise(
    minimum = min(TW_NET, na.rm = TRUE),
    maximum = max(TW_NET, na.rm = TRUE),
    mean = mean(TW_NET, na.rm = TRUE),
    median = median(TW_NET, na.rm = TRUE),
    sd = sd(TW_NET, na.rm = TRUE),
    cov = sd / mean * 100 # coefficient of variation
  )
```
= not much different!

## Largest Asset Europe

```{r}
df %>%
  filter(
    TIME_PERIOD == "2023-01-01",
    UNIT_MEASURE == "EUR",
    DWA_GRP == "_Z",
    REF_AREA == "I9",
    INSTR_ASSET %in% c("F51M", "NUB", "F62", "F52", "F511", "F3", "NUN", "F2M")
  ) %>%
  select(INSTR_ASSET, OBS_VALUE) %>%
  arrange(desc(OBS_VALUE)) %>%
  mutate(
    asset_share = OBS_VALUE / sum(OBS_VALUE) * 100, # calculate share of total assets
  )
```
